# model_api_item.go
/*
 * API Inventory
 *
 * This is the Challenge API
 *
 * API version: 1.0.0
 * Contact: maxmonte.vix@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
//package swagger
package main

import (
	"time"
)

type ApiItem struct {
	// API name
	Name string `bson:"name"`
	// API huma-readable name
	DisplayName string `bson:"display_name,omitempty"`
	// API detailed description
	Description string `bson:"description,omitempty"`
	// API  creation time stamp
	CreateTime time.Time `bson:"create_time,omitempty"`
	// API update time stamo
	UpdateTime time.Time `bson:"update_time,omitempty"`
	// openApi file name
	OpenapiFile string `bson:"openapi_file,omitempty"`
}